Refactor away extension traits RegionEscape and HasTypeFlags
