+++ src/librustc/middle/ty/fold.rs
@@ -42,7 +42,7 @@
 use middle::region;
 use middle::subst;
 use middle::ty::adjustment;
-use middle::ty::{self, Binder, Ty, RegionEscape, HasTypeFlags};
+use middle::ty::{self, Binder, Ty, TypeFlags};
 
 use std::fmt;
 use util::nodemap::{FnvHashMap, FnvHashSet};
@@ -58,6 +58,53 @@ pub trait TypeFoldable<'tcx>: fmt::Debug + Clone {
     fn visit_with<V: TypeVisitor<'tcx>>(&self, visitor: &mut V) -> bool;
     fn visit_subitems_with<V: TypeVisitor<'tcx>>(&self, visitor: &mut V) -> bool {
         self.visit_with(visitor)
+    }
+
+    fn has_regions_escaping_depth(&self, depth: u32) -> bool {
+        self.visit_with(&mut HasEscapingRegionsVisitor { depth: depth })
+    }
+    fn has_escaping_regions(&self) -> bool {
+        self.has_regions_escaping_depth(0)
+    }
+
+    fn has_type_flags(&self, flags: TypeFlags) -> bool {
+        self.visit_with(&mut HasTypeFlagsVisitor { flags: flags })
+    }
+    fn has_projection_types(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_PROJECTION)
+    }
+    fn references_error(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_TY_ERR)
+    }
+    fn has_param_types(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_PARAMS)
+    }
+    fn has_self_ty(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_SELF)
+    }
+    fn has_infer_types(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_TY_INFER)
+    }
+    fn needs_infer(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_TY_INFER | TypeFlags::HAS_RE_INFER)
+    }
+    fn needs_subst(&self) -> bool {
+        self.has_type_flags(TypeFlags::NEEDS_SUBST)
+    }
+    fn has_closure_types(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_TY_CLOSURE)
+    }
+    fn has_erasable_regions(&self) -> bool {
+        self.has_type_flags(TypeFlags::HAS_RE_EARLY_BOUND |
+                            TypeFlags::HAS_RE_INFER |
+                            TypeFlags::HAS_FREE_REGIONS)
+    }
+    /// Indicates whether this value references only 'global'
+    /// types/lifetimes that are the same regardless of what fn we are
+    /// in. This is used for caching. Errs on the side of returning
+    /// false.
+    fn is_global(&self) -> bool {
+        !self.has_type_flags(TypeFlags::HAS_LOCAL_NAMES)
     }
 }
 
@@ -518,64 +565,74 @@ pub fn shift_regions<'tcx, T:TypeFoldable<'tcx>>(tcx: &ty::ctxt<'tcx>,
     }))
 }
 
-impl<'tcx, T: TypeFoldable<'tcx>> RegionEscape for T {
-    fn has_regions_escaping_depth(&self, depth: u32) -> bool {
-        struct RegionEscapeVisitor {
-            depth: u32,
-        }
+/// An "escaping region" is a bound region whose binder is not part of `t`.
+///
+/// So, for example, consider a type like the following, which has two binders:
+///
+///    for<'a> fn(x: for<'b> fn(&'a isize, &'b isize))
+///    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ outer scope
+///                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~  inner scope
+///
+/// This type has *bound regions* (`'a`, `'b`), but it does not have escaping regions, because the
+/// binders of both `'a` and `'b` are part of the type itself. However, if we consider the *inner
+/// fn type*, that type has an escaping region: `'a`.
+///
+/// Note that what I'm calling an "escaping region" is often just called a "free region". However,
+/// we already use the term "free region". It refers to the regions that we use to represent bound
+/// regions on a fn definition while we are typechecking its body.
+///
+/// To clarify, conceptually there is no particular difference between an "escaping" region and a
+/// "free" region. However, there is a big difference in practice. Basically, when "entering" a
+/// binding level, one is generally required to do some sort of processing to a bound region, such
+/// as replacing it with a fresh/skolemized region, or making an entry in the environment to
+/// represent the scope to which it is attached, etc. An escaping region represents a bound region
+/// for which this processing has not yet been done.
+struct HasEscapingRegionsVisitor {
+    depth: u32,
+}
 
-        impl<'tcx> TypeVisitor<'tcx> for RegionEscapeVisitor {
-            fn enter_region_binder(&mut self) {
-                self.depth += 1;
-            }
+impl<'tcx> TypeVisitor<'tcx> for HasEscapingRegionsVisitor {
+    fn enter_region_binder(&mut self) {
+        self.depth += 1;
+    }
 
-            fn exit_region_binder(&mut self) {
-                self.depth -= 1;
-            }
+    fn exit_region_binder(&mut self) {
+        self.depth -= 1;
+    }
 
-            fn visit_ty(&mut self, t: Ty<'tcx>) -> bool {
-                t.region_depth > self.depth
-            }
+    fn visit_ty(&mut self, t: Ty<'tcx>) -> bool {
+        t.region_depth > self.depth
+    }
 
-            fn visit_region(&mut self, r: ty::Region) -> bool {
-                r.escapes_depth(self.depth)
-            }
-        }
-
-        self.visit_with(&mut RegionEscapeVisitor { depth: depth })
+    fn visit_region(&mut self, r: ty::Region) -> bool {
+        r.escapes_depth(self.depth)
     }
 }
 
-impl<'tcx, T: TypeFoldable<'tcx>> HasTypeFlags for T {
-    fn has_type_flags(&self, flags: ty::TypeFlags) -> bool {
-        struct HasTypeFlagsVisitor {
-            flags: ty::TypeFlags,
-        }
+struct HasTypeFlagsVisitor {
+    flags: ty::TypeFlags,
+}
 
-        impl<'tcx> TypeVisitor<'tcx> for HasTypeFlagsVisitor {
-            fn visit_ty(&mut self, t: Ty) -> bool {
-                t.flags.get().intersects(self.flags)
-            }
+impl<'tcx> TypeVisitor<'tcx> for HasTypeFlagsVisitor {
+    fn visit_ty(&mut self, t: Ty) -> bool {
+        t.flags.get().intersects(self.flags)
+    }
 
-            fn visit_region(&mut self, r: ty::Region) -> bool {
-                if self.flags.intersects(ty::TypeFlags::HAS_LOCAL_NAMES) {
-                    // does this represent a region that cannot be named
-                    // in a global way? used in fulfillment caching.
-                    match r {
-                        ty::ReStatic | ty::ReEmpty => {}
-                        _ => return true,
-                    }
-                }
-                if self.flags.intersects(ty::TypeFlags::HAS_RE_INFER) {
-                    match r {
-                        ty::ReVar(_) | ty::ReSkolemized(..) => { return true }
-                        _ => {}
-                    }
-                }
-                false
+    fn visit_region(&mut self, r: ty::Region) -> bool {
+        if self.flags.intersects(ty::TypeFlags::HAS_LOCAL_NAMES) {
+            // does this represent a region that cannot be named
+            // in a global way? used in fulfillment caching.
+            match r {
+                ty::ReStatic | ty::ReEmpty => {}
+                _ => return true,
             }
         }
-
-        self.visit_with(&mut HasTypeFlagsVisitor { flags: flags })
+        if self.flags.intersects(ty::TypeFlags::HAS_RE_INFER) {
+            match r {
+                ty::ReVar(_) | ty::ReSkolemized(..) => { return true }
+                _ => {}
+            }
+        }
+        false
     }
 }
