+++ src/librustc_trans/trans/meth.rs
@@ -34,7 +34,7 @@ use trans::machine;
 use trans::monomorphize;
 use trans::type_::Type;
 use trans::type_of::*;
-use middle::ty::{self, Ty, HasTypeFlags};
+use middle::ty::{self, Ty, TypeFoldable};
 use middle::ty::MethodCall;
 
 use syntax::ast;
